First running int main() {return 0;} 1) Fixed inline parser 2) Included AsmParser.inc

+++ mytest/hello/hello.c
@@ -1,3 +1,11 @@
-int main(){
-	return 0;
+//int a = 10;
+//int f() {
+//    return a;
+//}
+//int a = 10, b = 20, c;
+long main() {
+    //c = a + b;
+    //return f();
+    return 0;
 }
+
